Use atomic loads for flags map
Previously the map was only read by All and Lookup, which are rarely usde.

Because of that we used a simple mutex to guard access to the flags map.

With the switch from registries to context-specific flag values, this is no
longer true, as setting the values requires going over the flags.

This change moves the map behind an atomic value that can be read without a
lock, which should remove all potential contention.